fix: test useDevice
