+++ sections/Product/ProductShelfWithComparatorSection.tsx
@@ -1,7 +1,7 @@
 import ProductShelf from "site/islands/ProductShelf.tsx";
 import Section from "site/components/ui/Section.tsx";
 import { useId } from "site/sdk/useId.ts";
-import { useDevice } from "@deco/deco/hooks";
+import { useIsDesktop } from "site/hooks/useIsDesktop.ts";
 import type { ISection } from "site/types/Section.d.ts";
 import type { ISliderConfigs } from "site/types/Slider.d.ts";
 import type { ProductSpecsComparator } from "site/types/Product.d.ts";
@@ -56,7 +56,7 @@ export default function ProductShelfWithComparatorSection({
   products
 }: Props): preact.JSX.Element {
   const id = useId();
-  const device = useDevice();
+  const isDesktop = useIsDesktop();
 
   const hasProducts = !!products?.length;
 
@@ -77,10 +77,10 @@ export default function ProductShelfWithComparatorSection({
     slidesPerView: 1.5,
     spaceBetween: 8,
     pagination: {
-      enabled: device === "desktop" ? configs?.pagination?.enabledDesktop : configs?.pagination?.enabledMobile
+      enabled: isDesktop.value ? configs?.pagination?.enabledDesktop : configs?.pagination?.enabledMobile
     },
     navigation: {
-      enabled: device === "desktop" ? configs?.navigation?.enabledDesktop : configs?.navigation?.enabledMobile
+      enabled: isDesktop.value ? configs?.navigation?.enabledDesktop : configs?.navigation?.enabledMobile
     },
     breakpoints: {
       768: {
